+++ R/svyjskm.R
@@ -12,6 +12,8 @@
 #' @param timeby numeric: control the granularity along the time-axis; defaults to 7 time-points. 
 #' @param main plot title, Default: ''
 #' @param pval logical: add the pvalue to the plot?, Default: FALSE
+#' @param pval.size numeric value specifying the p-value text size. Default is 5.
+#' @param pval.coord numeric vector, of length 2, specifying the x and y coordinates of the p-value. Default values are NULL
 #' @param legend logical. should a legend be added to the plot? Default: TRUE
 #' @param ci logical. Should confidence intervals be plotted. Default = NULL
 #' @param legendposition numeric. x, y position of the legend if plotted. Default: c(0.85, 0.8)
@@ -54,6 +56,8 @@ svyjskm <- function(sfit,
                     timeby = NULL,
                     main = "",
                     pval = FALSE,
+                    pval.size = 5, 
+                    pval.coord = c(NULL, NULL),
                     legend = TRUE,
                     legendposition=c(0.85,0.8),
                     ci = NULL,
@@ -217,7 +221,11 @@ svyjskm <- function(sfit,
     
     pvaltxt <- ifelse(pvalue < 0.0001,"p < 0.0001",paste("p =", signif(pvalue, 3)))
     # MOVE P-VALUE LEGEND HERE BELOW [set x and y]
-    p <- p + annotate("text",x = (as.integer(max(sapply(sfit, function(x){max(x$time)}))/5)), y = 0.1 + ylims[1],label = pvaltxt)
+    if (is.null(pval.coord)){
+      p <- p + annotate("text",x = (as.integer(max(sfit$time)/5)), y = 0.1 + ylims[1],label = pvaltxt, size  = pval.size)
+    } else{
+      p <- p + annotate("text",x = pval.coord[1], y = pval.coord[2], label = pvaltxt, size  = pval.size)
+    }
   }
   
   p