+++ R/jskm.R
@@ -278,12 +278,12 @@ jskm <- function(sfit,
     pvalue <- pchisq(sdiff$chisq,length(sdiff$n) - 1,lower.tail = FALSE)
     
     ## cluster option
-    if (cluster.option == "cluster" && !is.null(cluster.var)){
+    if (cluster.option == "cluster" & !is.null(cluster.var)){
       form.old = as.character(sfit$call$formula)
       form.new = paste(form.old[2], form.old[1], " + ", form.old[3], " + cluster(", cluster.var, ")", sep="")
       sdiff <- survival::coxph(as.formula(form.new), data = data)
       pvalue <- summary(sdiff)$robscore["pvalue"]
-    } else if (cluster.option == "frailty" && !is.null(cluster.var)){
+    } else if (cluster.option == "frailty" & !is.null(cluster.var)){
       form.old = as.character(sfit$call$formula)
       form.new = paste(form.old[2], form.old[1], " + ", form.old[3], " + frailty(", cluster.var, ")", sep="")
       sdiff <- survival::coxph(as.formula(form.new), data =data)